Skip BinLog Tests on MS Windows Platforms
Add null check on cleanup() as there are no files in the Temp Dir since we are skipping the test.

+++ modules/binlog/src/test/java/org/jpos/binlog/BinLogTest.java
@@ -89,10 +89,12 @@ public class BinLogTest implements Runnable {
 
     @AfterClass
     public static void cleanup() throws IOException {
-        for (File f : dir.listFiles()) {
-            if (f.toString().endsWith(".dat")) {
-                System.out.println ("Deleting " + f.toString());
-                f.delete();
+        if (dir.listFiles() != null) {
+            for (File f : dir.listFiles()) {
+                if (f.toString().endsWith(".dat")) {
+                    System.out.println ("Deleting " + f.toString());
+                    f.delete();
+                }
             }
         }
         System.out.println ("Deleting " + dir);
